# apps/api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Omnibase REST API
  version: 0.9.18
  description: |
    Self-hostable Backend-as-a-Service providing database management, authentication, payments, storage, and email services.

    ## Features
    - **Database**: PostgreSQL with RLS and migrations
    - **Authentication**: Ory Kratos integration with session management
    - **Payments**: Stripe integration with version-controlled billing configs
    - **Storage**: S3-compatible object storage with RLS
    - **Email**: Transactional email service
    - **Permissions**: Fine-grained access control via Ory Keto

    ## Authentication
    Most endpoints require authentication via session cookies or JWT tokens.
    Use the appropriate security scheme based on the endpoint requirements.
  contact:
    name: Omnibase Support
    url: https://omnibase.dev/support
    email: support@omnibase.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://omnibase.dev/terms
servers:
  - url: https://api.omnibase.tech
    description: Production server
...
paths:
...
  /api/v1/permissions/check:
    post:
      operationId: checkPermission
      summary: Check permission
      description: |
        Checks if a subject has a specific permission on an object using Ory Keto.

        ## Authentication
        Requires session authentication.

        ## Request Format
        Provide either `subject_id` or `subject_set` (not both). The request uses oneOf 
        to enforce this constraint at the schema level.

        ## Use Cases
        - Verify user permissions before performing actions
        - Implement fine-grained access control
        - Check role-based permissions
      tags:
        - V1 Permissions
      security:
        - CookieAuth: []
          SessionTokenAuth: []
          ServiceKeyAuth: []
      requestBody:
        required: true
        description: Permission check request with either subject_id or subject_set
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckPermissionRequest'
            examples:
              withSubjectId:
                summary: Check permission with subject_id
                value:
                  namespace: Tenant
                  object: tenant_test_123
                  relation: can_invite_user
                  subject_id: 550e8400-e29b-41d4-a716-446655440000
              withSubjectSet:
                summary: Check permission with subject_set
                value:
                  namespace: Tenant
                  object: tenant_test_123
                  relation: can_delete_tenant
                  subject_set:
                    namespace: Tenant
                    object: tenant_parent_456
                    relation: can_invite_user
      responses:
        '200':
          description: Permission check result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CheckPermissionResponse'
              examples:
                allowed:
                  summary: Permission granted
                  value:
                    status: 200
                    data:
                      allowed: true
                denied:
                  summary: Permission denied
                  value:
                    status: 200
                    data:
                      allowed: false
        '400':
          description: Invalid request body - namespace, object, and relation cannot be empty, or both/neither subject fields provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
              examples:
                bothSubjects:
                  summary: Both subject_id and subject_set provided
                  value:
                    status: 400
                    error: Provide either subject_id or subject_set, not both
                neitherSubject:
                  summary: Neither subject_id nor subject_set provided
                  value:
                    status: 400
                    error: Either subject_id or subject_set must be provided
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        '500':
          description: Failed to check permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorResponse'
  /api/v1/permissions/relationships:
    delete:
      operationId: deleteRelationship
      summary: Delete relationship
      description: |
        Deletes a relationship tuple from Ory Keto.

        ## Authentication
        Requires session authentication.

        ## Request Format
        Provide either `subject_id` or `subject_set` (not both). The request uses oneOf
        to enforce this constraint at the schema level.

        ## Use Cases
        - Remove resource links from tenants
        - Revoke user assignments from projects
        - Delete permission relationships
      tags:
        - V1 Permissions
      security:
        - CookieAuth: []
          SessionTokenAuth: []
        - ServiceKeyAuth: []
      requestBody:
        required: true
        description: Relationship deletion request with either subject_id or subject_set
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteRelationshipRequest'
            examples:
              withSubjectId:
                summary: Delete relationship with subject_id
                value:
                  namespace: Tenant
                  object: tenant_test_123
                  relation: can_invite_user
                  subject_id: 550e8400-e29b-41d4-a716-446655440000
              withSubjectSet:
                summary: Delete relationship with subject_set
                value:
                  namespace: Project
                  object: project_123
                  relation: tenant
                  subject_set:
                    namespace: Tenant
                    object: tenant_456
      responses:
        '200':
          description: Relationship deleted successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/DeleteRelationshipResponse'
              example:
                status: 200
                data:
                  message: Relationship deleted successfully
        '400':
          description: Invalid request body - namespace, object, and relation are required, or both/neither subject fields provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
              examples:
                bothSubjects:
                  summary: Both subject_id and subject_set provided
                  value:
                    status: 400
                    error: Provide either subject_id or subject_set, not both
                neitherSubject:
                  summary: Neither subject_id nor subject_set provided
                  value:
                    status: 400
                    error: Either subject_id or subject_set must be provided
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          description: Failed to delete relationship
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorResponse'
    post:
      operationId: createRelationship
      summary: Create relationship
      description: |
        Creates a new relationship tuple in Ory Keto.

        ## Authentication
        Requires session authentication.

        ## Request Format
        Provide either `subject_id` or `subject_set` (not both). The request uses oneOf 
        to enforce this constraint at the schema level.

        ## Use Cases
        - Link resources to tenants
        - Assign users to projects
        - Create permission relationships
      tags:
        - V1 Permissions
      security:
        - CookieAuth: []
          SessionTokenAuth: []
        - ServiceKeyAuth: []
      requestBody:
        required: true
        description: Relationship creation request with either subject_id or subject_set
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRelationshipRequest'
            examples:
              withSubjectId:
                summary: Create relationship with subject_id
                value:
                  namespace: Tenant
                  object: tenant_test_123
                  relation: can_invite_user
                  subject_id: 550e8400-e29b-41d4-a716-446655440000
              withSubjectSet:
                summary: Create relationship with subject_set
                value:
                  namespace: Tenant
                  object: tenant_test_123
                  relation: can_delete_tenant
                  subject_set:
                    namespace: Tenant
                    object: tenant_parent_456
                    relation: can_invite_user
      responses:
        '200':
          description: Relationship created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CreateRelationshipResponse'
              example:
                status: 200
                data:
                  message: Relationship created successfully
                  relationship:
                    namespace: Tenant
                    object: tenant_test_123
                    relation: can_invite_user
                    subject_id: 550e8400-e29b-41d4-a716-446655440000
        '400':
          description: Invalid request body - namespace, object, and relation are required, or both/neither subject fields provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
              examples:
                bothSubjects:
                  summary: Both subject_id and subject_set provided
                  value:
                    status: 400
                    error: Provide either subject_id or subject_set, not both
                neitherSubject:
                  summary: Neither subject_id nor subject_set provided
                  value:
                    status: 400
                    error: Either subject_id or subject_set must be provided
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          description: Failed to create relationship
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorResponse'
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: ory_kratos_session
      description: |
        Session cookie authentication. Cookie name: `ory_kratos_session`.
        Automatically set by browser after Kratos login.
    SessionTokenAuth:
      type: apiKey
      in: header
      name: X-Session-Token
      description: |
        Kratos session JWT token. Alternative to cookie authentication for
        non-browser clients. Obtain from Kratos after login.
    ServiceKeyAuth:
      type: apiKey
      in: header
      name: X-Service-Key
      description: |
        Service-to-service authentication key for backend operations. When
        used with tenant endpoints, must include X-Tenant-ID header.
  schemas:
    SuccessResponse:
      type: object
      description: Standard success response wrapper
      properties:
        status:
          type: integer
          description: HTTP status code
          example: 200
        data:
          description: Response data payload
          oneOf:
            - type: object
            - type: array
            - type: string
            - type: number
            - type: boolean
      required:
        - status
...
    BadRequestResponse:
      type: object
      description: Bad Request error response (400)
      properties:
        status:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: Error message or details
          example: Invalid request parameters
      required:
        - status
        - error
    UnauthorizedResponse:
      type: object
      description: Unauthorized error response (401)
      properties:
        status:
          type: integer
          description: HTTP status code
          example: 401
        error:
          type: string
          description: Error message
          example: Unauthorized
      required:
        - status
        - error
...
    InternalServerErrorResponse:
      type: object
      description: Internal Server Error response (500)
      properties:
        status:
          type: integer
          description: HTTP status code
          example: 500
        error:
          type: string
          description: Error message
          example: Internal Server Error
      required:
        - status
        - error
...
    CheckPermissionRequestWithSubjectId:
      type: object
      description: Check permission request using a direct subject identifier
      required:
        - namespace
        - object
        - relation
        - subject_id
      properties:
        namespace:
          type: string
          minLength: 1
          description: The namespace of the permission check
        object:
          type: string
          minLength: 1
          description: The object to check permissions on
        relation:
          type: string
          minLength: 1
          description: The relation/permission to check
        subject_id:
          type: string
          description: Direct subject identifier
      additionalProperties: false
...
    CheckPermissionRequestWithSubjectSet:
      type: object
      description: Check permission request using a subject set
      required:
        - namespace
        - object
        - relation
        - subject_set
      properties:
        namespace:
          type: string
          minLength: 1
          description: The namespace of the permission check
        object:
          type: string
          minLength: 1
          description: The object to check permissions on
        relation:
          type: string
          minLength: 1
          description: The relation/permission to check
        subject_set:
          $ref: '#/components/schemas/SubjectSetRequest'
      additionalProperties: false
    CheckPermissionRequest:
      description: Check permission request. Must provide exactly one of subject_id or subject_set.
      oneOf:
        - $ref: '#/components/schemas/CheckPermissionRequestWithSubjectId'
        - $ref: '#/components/schemas/CheckPermissionRequestWithSubjectSet'
    CheckPermissionResponse:
      type: object
      description: Permission check result
      properties:
        allowed:
          type: boolean
          description: Whether the permission check passed
          example: true
      required:
        - allowed
...
    CreateRelationshipRequest:
      description: Create relationship request. Must provide exactly one of subject_id or subject_set.
      oneOf:
        - $ref: '#/components/schemas/CreateRelationshipRequestWithSubjectId'
        - $ref: '#/components/schemas/CreateRelationshipRequestWithSubjectSet'
...
    CreateRelationshipResponse:
      type: object
      description: Relationship creation result
      properties:
        message:
          type: string
          description: Success message
          example: Relationship created successfully
        relationship:
          $ref: '#/components/schemas/Relationship'
      required:
        - message
        - relationship
    DeleteRelationshipRequestWithSubjectId:
      type: object
      description: Delete relationship request using a direct subject identifier
      required:
        - namespace
        - object
        - relation
        - subject_id
      properties:
        namespace:
          type: string
          description: The namespace for the relationship
        object:
          type: string
          description: The object in the relationship
        relation:
          type: string
          description: The relation type
        subject_id:
          type: string
          description: Direct subject identifier
      additionalProperties: false
    DeleteRelationshipRequestWithSubjectSet:
      type: object
      description: Delete relationship request using a subject set
      required:
        - namespace
        - object
        - relation
        - subject_set
      properties:
        namespace:
          type: string
          description: The namespace for the relationship
        object:
          type: string
          description: The object in the relationship
        relation:
          type: string
          description: The relation type
        subject_set:
          $ref: '#/components/schemas/SubjectSetRequest'
      additionalProperties: false
    DeleteRelationshipRequest:
      description: Delete relationship request. Must provide exactly one of subject_id or subject_set.
      oneOf:
        - $ref: '#/components/schemas/DeleteRelationshipRequestWithSubjectId'
        - $ref: '#/components/schemas/DeleteRelationshipRequestWithSubjectSet'
    DeleteRelationshipResponse:
      type: object
      description: Relationship deletion result
      properties:
        message:
          type: string
          description: Success message
          example: Relationship deleted successfully
      required:
        - message
...
  responses:
...
    NotFoundError:
      description: Not Found - Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotFoundResponse'
...